code/12_paketierung/requirementsx.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
chatbotAI==0.3.0.0
click==7.1.2
colorama==0.4.4
comtypes==1.1.7
cx-Freeze==6.5.3
decorator==4.4.2
deprecation==2.1.0
docopt==0.6.2
enum34==1.1.10
ffmpeg-python==0.2.0
future==0.17.1
fuzzywuzzy==0.18.0
geocoder==1.38.1
geojson==2.5.0
idna==2.10
importlib-metadata==3.10.0
joblib==0.17.0
loguru==0.5.3
notify-py @ git+https://github.com/ms7m/notify-py/
num2words==0.5.10
numpy==1.19.3
packaging==20.4
Pillow==8.0.1
pvporcupine==1.8.7
pyaml==19.12.0
PyAudio-0.2.11-cp38-cp38-win_amd64.whl
pycountry==20.7.3
pycparser==2.20
pyowm==3.1.1
pyparsing==2.4.7
pypiwin32==223
PySocks==1.7.1
python-crfsuite==0.9.7
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2020.4
pywin32==228
PyYAML==5.3.1
ratelim==0.1.6
requests==2.24.0
scikit-learn==0.22.2.post1
scipy==1.5.3
six==1.15.0
sklearn-crfsuite==0.3.6
snips-nlu==0.20.2
snips-nlu-de @ https://github.com/snipsco/snips-nlu-language-resources/releases/download/snips_nlu_de-0.2.3/snips_nlu_de-0.2.3.tar.gz
snips-nlu-parsers==0.4.3
snips-nlu-utils==0.9.1
sounddevice==0.4.1
SoundFile==0.10.2
soupsieve==2.1
tabulate==0.8.7
text2numde==1.0
tinydb==4.2.0
tqdm==4.51.0
urllib3==1.25.11
vosk==0.3.7
wikipedia==1.4.0
win10toast==0.9
win32-setctime==1.0.3
wincertstore==0.2
wxPython==4.1.1
zipp==3.4.1
